pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentbench"
version = "0.1.0"
description = "An eval framework for agentic coding tools"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "Anupam Mediratta", email = "mediratta@gmail.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "typer>=0.12,<1.0",
    "rich>=13.0,<14.0",
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "docker>=7.0,<8.0",
    "anthropic[bedrock]>=0.40,<1.0",
    "aiohttp>=3.9,<4.0",
    "jinja2>=3.1,<4.0",
    "jsonschema>=4.20,<5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "pytest-cov>=5.0,<6.0",
    "ruff>=0.5,<1.0",
    "mypy>=1.10,<2.0",
    "pre-commit>=3.7,<4.0",
]

[project.urls]
Homepage = "https://github.com/anupamme/AgentBench"
Repository = "https://github.com/anupamme/AgentBench"
Documentation = "https://github.com/anupamme/AgentBench#readme"
Issues = "https://github.com/anupamme/AgentBench/issues"


[project.scripts]
agentbench = "agentbench.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/agentbench"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.per-file-ignores]
"src/agentbench/reporting/reporter.py" = ["E501"]  # Jinja2 template rows can't be wrapped
"tests/test_reporting/*.py" = ["E501"]              # Test fixture dicts are intentionally verbose

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = ["docker: tests requiring Docker daemon"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
mypy_path = "src"
disable_error_code = ["misc"]  # Typer decorators are intentionally untyped
